refactor(api): migrate controller and serializer for /api/admin/users/{id}/organizations
