+++ api/src/team/application/membership/membership.admin.controller.js
@@ -1,5 +1,6 @@
 import * as membershipSerializer from '../../../shared/infrastructure/serializers/jsonapi/membership.serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
+import * as userOrganizationForAdminSerializer from '../../infrastructure/serializers/jsonapi/user-organization-for-admin-serializer.js';
 
 const findPaginatedFilteredMembershipsForAdmin = async function (request) {
   const organizationId = request.params.id;
@@ -13,8 +14,18 @@ const findPaginatedFilteredMembershipsForAdmin = async function (request) {
   return membershipSerializer.serializeForAdmin(memberships, pagination);
 };
 
-const membershipAdminController = {
-  findPaginatedFilteredMembershipsForAdmin,
+const findUserOrganizationsForAdmin = async function (
+  request,
+  h,
+  dependencies = { userOrganizationForAdminSerializer },
+) {
+  const userId = request.params.id;
+  const organizations = await usecases.findUserOrganizationsForAdmin({ userId });
+
+  return h.response(dependencies.userOrganizationForAdminSerializer.serialize(organizations));
 };
 
-export { membershipAdminController };
+export const membershipAdminController = {
+  findPaginatedFilteredMembershipsForAdmin,
+  findUserOrganizationsForAdmin,
+};
